+++ cypress.env.json
@@ -1,3 +1,4 @@
 {
-    "password":"password"
+    "password":"password",
+    "user" : "test"
 }
